add default decoder for other disputes (#23)
* add default decoder for other disputes

* remove EOFS
