+++ src/components/Disputes/DisputeActions.js
@@ -58,6 +58,14 @@ function DisputeActions({
 
   const guardianHasVoted = isGuardianDrafted && hasGuardianVoted(guardianDraft)
 
+  // prepare button appearences.
+  const votingButtons = {
+    allowActionText: dispute.metadata?.allowActionText,
+    allowActionColor: dispute.metadata?.allowActionColor,
+    blockActionText: dispute.metadata?.blockActionText,
+    blockActionColor: dispute.metadata?.blockActionColor
+  }
+
   if (phase === DisputePhase.VotingPeriod && !guardianHasVoted) {
     return (
       <DisputeVoting
@@ -65,6 +73,7 @@ function DisputeActions({
         isFinalRound={dispute.maxAppealReached}
         isGuardianDrafted={isGuardianDrafted}
         onRequestCommit={onRequestCommit}
+        buttons={votingButtons}
       />
     )
   }
