+++ src/utils/dispute-metadata.js
@@ -1,8 +1,20 @@
 import { decode as GovernDisputeDecode } from './metadata/GovernDispute'
+import { decode as DefaultDecoder } from './metadata/DefaultDispute';
+import { DISPUTE_TYPES } from './metadata/types';
 
-export  async function decodeMetadata(metadata, party = 'Govern') {
-    if(party === 'Govern') {
-       const data =  await GovernDisputeDecode(metadata);
+export  async function decodeMetadata(metadata) {
+    let data = null;
+    
+    try {
+       data = await DefaultDecoder(metadata)
+       data.disputeType = DISPUTE_TYPES.DEFAULT;
        return data;
-    }
+    }catch(err) {}
+
+    try {
+        data =  await GovernDisputeDecode(metadata);
+        data.disputeType = DISPUTE_TYPES.GOVERN;
+    }catch(err) {}
+    
+    return data;
 }
