+++ src/utils/metadata/types.js
@@ -1,0 +1,5 @@
+// Can be used Symbol as well(much better)
+export const DISPUTE_TYPES = {
+    DEFAULT: 'DEFAULT',
+    GOVERN: 'GOVERN'
+}
